requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlmodel
sqlalchemy
alembic

python-multipart

passlib[bcrypt]
python-jose[cryptography]

pydantic
pydantic-settings
